make deserialize optional
