+++ src/record/field.rs
@@ -4,9 +4,6 @@ use std::io::{Read, Write, Seek, SeekFrom};
 
 use std::str::FromStr;
 
-use serde::de::{Deserialize, Visitor};
-use serde::Deserializer;
-use serde::export::Formatter;
 use byteorder::{LittleEndian, BigEndian, ReadBytesExt, WriteBytesExt};
 
 use record::FieldInfo;
@@ -222,24 +219,32 @@ pub struct Date {
     day: u32,
 }
 
-impl<'de> Deserialize<'de> for Date {
-    fn deserialize<D>(deserializer: D) -> Result<Self, <D as Deserializer<'de>>::Error> where
-        D: Deserializer<'de> {
-        struct DateVisitor;
-        impl<'de> Visitor<'de> for DateVisitor {
-            type Value = Date;
+#[cfg(feature = "serde")]
+mod de {
+    use super::*;
+    use serde::de::{Deserialize, Visitor};
+    use serde::Deserializer;
+    use serde::export::Formatter;
 
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("struct Date")
-            }
+    impl<'de> Deserialize<'de> for Date {
+        fn deserialize<D>(deserializer: D) -> Result<Self, <D as Deserializer<'de>>::Error> where
+            D: Deserializer<'de> {
+            struct DateVisitor;
+            impl<'de> Visitor<'de> for DateVisitor {
+                type Value = Date;
 
-            fn visit_byte_buf<E>(self, v: Vec<u8>) -> Result<Self::Value, E> where
-                E: serde::de::Error, {
-                let string = String::from_utf8(v).unwrap();
-                Ok(Date::from_str(&string).unwrap())
+                fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+                    formatter.write_str("struct Date")
+                }
+
+                fn visit_byte_buf<E>(self, v: Vec<u8>) -> Result<Self::Value, E> where
+                    E: serde::de::Error, {
+                    let string = String::from_utf8(v).unwrap();
+                    Ok(Date::from_str(&string).unwrap())
+                }
             }
+            deserializer.deserialize_byte_buf(DateVisitor)
         }
-        deserializer.deserialize_byte_buf(DateVisitor)
     }
 }
